feat: swap from SVG --> PNG, fix theme change
