+++ lib/assets.dart
@@ -2,6 +2,6 @@ class AppAssets {
   static const String PROTOCOL_2019 = 'wvems-protocols-2019';
   static const String PROTOCOL_2020 = 'wvems-protocols-2020';
 
-  static const String SVG_LOGO_DARK = 'assets/svg/wvems_logo_dark.svg';
-  static const String SVG_LOGO_LIGHT = 'assets/svg/wvems_logo_light.svg';
+  static const String PNG_LOGO_DARK = 'assets/png/wvems_logo_dark.png';
+  static const String PNG_LOGO_LIGHT = 'assets/png/wvems_logo_light.png';
 }
