+++ lib/ui/styled_components/styled_components.dart
@@ -1,3 +1,4 @@
 export 'app_logo.dart';
 export 'styled_dialog.dart';
 export 'styled_protocols_year.dart';
+export 'styled_ribbon_stack.dart';
