+++ lib/ui/views/nav_drawer/nav_header.dart
@@ -1,7 +1,11 @@
 import 'package:flutter/material.dart';
+import 'package:gap/gap.dart';
+import 'package:wvems_protocols/_internal/utils/utils.dart';
 import 'package:wvems_protocols/ui/styled_components/styled_components.dart';
 
 class NavHeader extends StatelessWidget {
+  final String _yearText = wvemsText(2020);
+
   @override
   Widget build(BuildContext context) {
     // being displayed. The color should be linked to the year, just like
@@ -10,14 +14,15 @@ class NavHeader extends StatelessWidget {
     // And _displayWompWomp() is just a placeholder so the menus work.
     // <kludge>
     return Padding(
-      padding: const EdgeInsets.all(12.0),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.stretch,
+      padding: const EdgeInsets.all(4.0),
+      child: StyledRibbonStack(
+        title: Text(_yearText),
+        padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 4),
+        hasOkButton: false,
         children: <Widget>[
+          const Gap(30),
           DrawerAppLogo(),
-          const SizedBox(height: 8.0),
-          StyledProtocolsYear(),
+          const Gap(30),
         ],
       ),
     );
