+++ pubspec.yaml
@@ -30,7 +30,6 @@ dependencies:
     sdk: flutter
   flutter_local_notifications: ^5.0.0
   flutter_pdfview: ^1.1.0
-  flutter_svg: ^0.21.0-nullsafety.0
   freezed: ^0.14.1+1
   gap: 2.0.0-nullsafety.1
   get: ^4.1.1
@@ -64,8 +63,8 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/json/
+    - assets/png/
     - assets/pdf/
-    - assets/svg/
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
   # For details regarding adding assets from package dependencies, see
